requirements.txt
matplotlib==3.4.2
numpy==1.19.5
pandas==1.3.3
Pillow==8.4.0
plotly==5.3.1
scikit_learn==1.0.1
tensorflow==2.6.0
